Refactor ingress-nginx conversion logic
